+++ processor/src/test/resources/bad_input/com/example/inject/MultipleConstructorsModel.java
@@ -1,0 +1,21 @@
+package com.example.inject;
+
+import org.realityforge.arez.annotations.Action;
+import org.realityforge.arez.annotations.ArezComponent;
+
+@ArezComponent( inject = true )
+public class MultipleConstructorsModel
+{
+  public MultipleConstructorsModel()
+  {
+  }
+
+  public MultipleConstructorsModel( int i )
+  {
+  }
+
+  @Action
+  void myAction()
+  {
+  }
+}