+++ processor/src/test/resources/input/com/example/inject/BasicInjectModel.java
@@ -1,0 +1,17 @@
+package com.example.inject;
+
+import org.realityforge.arez.annotations.Action;
+import org.realityforge.arez.annotations.ArezComponent;
+
+@ArezComponent( inject = true )
+public class BasicInjectModel
+{
+  public BasicInjectModel()
+  {
+  }
+
+  @Action
+  public void myActionStuff()
+  {
+  }
+}