+++ processor/src/test/resources/input/com/example/inject/DefaultCtorModel.java
@@ -1,0 +1,13 @@
+package com.example.inject;
+
+import org.realityforge.arez.annotations.Action;
+import org.realityforge.arez.annotations.ArezComponent;
+
+@ArezComponent( inject = true )
+public class DefaultCtorModel
+{
+  @Action
+  public void myActionStuff()
+  {
+  }
+}