+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   effectively removes the `dispose()` method from the public interface of Observers accessed solely through
   the spy subsystem.
 * **\[core\]** Implement `equals()` and `hashCode()` on `ComponentInfoImpl`.
+* **\[annotations\]** Add an `inject` parameter to `@ArezComponent` annotation that will add a
+  `@javax.inject.Inject` annotation on the generated classes constructor if set to true.
 
 ### [v0.30](https://github.com/arez/arez/tree/v0.30) (2017-11-29)
 [Full Changelog](https://github.com/arez/arez/compare/v0.29...v0.30)